+++ src/main/java/com/example/demo/user/UserRepository.java
@@ -1,4 +1,4 @@
-package com.example.demo.entity;
+package com.example.demo.user;
 
 import org.springframework.data.repository.CrudRepository;
