+++ src/main/java/com/example/demo/user/UserService.java
@@ -1,4 +1,4 @@
-package com.example.demo.entity;
+package com.example.demo.user;
 
 import org.springframework.beans.factory.annotation.Autowired;
